+++ src/PackageSkeleton.php
@@ -1,15 +1,15 @@
 <?php
 
 /*
- * This file is part of the guanguans/di.
+ * This file is part of the guanguans/package-skeleton.
  *
  * (c) guanguans <ityaozm@gmail.com>
  *
  * This source file is subject to the MIT license that is bundled.
  */
 
-namespace Guanguans\Di;
+namespace Guanguans\PackageSkeleton;
 
-class Di
+class PackageSkeleton
 {
 }