+++ tests/PackageSkeletonTest.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of the guanguans/package-skeleton.
+ *
+ * (c) guanguans <ityaozm@gmail.com>
+ *
+ * This source file is subject to the MIT license that is bundled.
+ */
+
+namespace Guanguans\PackageSkeleton\Tests;
+
+class PackageSkeletonTest extends TestCase
+{
+    public function testTest()
+    {
+        $this->assertIsString('This is a test.');
+    }
+}