+++ tests/TestCase.php
@@ -1,14 +1,14 @@
 <?php
 
 /*
- * This file is part of the guanguans/di.
+ * This file is part of the guanguans/package-skeleton.
  *
  * (c) guanguans <ityaozm@gmail.com>
  *
  * This source file is subject to the MIT license that is bundled.
  */
 
-namespace Guanguans\Di\Tests;
+namespace Guanguans\PackageSkeleton\Tests;
 
 class TestCase extends \PHPUnit\Framework\TestCase
 {